backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
DateTime==5.5
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
SQLAlchemy==2.0.34
SQLAlchemy-Utils==0.41.2
tabulate==0.9.0
typing_extensions==4.12.2
Werkzeug==3.0.4
zope.interface==7.0.3
